% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{check_lock_support_webdav}
\alias{check_lock_support_webdav}
\title{Check if a resource allows locking using PROPFIND}
\usage{
check_lock_support_webdav(
  base_url,
  resource_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD")
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server.}

\item{resource_path}{The path of the resource to check.}

\item{dav}{The WebDAV base path.}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}
}
\value{
Logical value indicating whether the resource supports locking.
}
\description{
This function checks if the resource allows the LOCK method by retrieving properties with PROPFIND.
}
