% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_response}
\alias{handle_response}
\title{Handle HTTP response from Server}
\usage{
handle_response(response)
}
\arguments{
\item{response}{The response object from an `httr2` request.}
}
\value{
The processed response object if successful, or an error if the request failed.
}
\description{
This function processes the response from the WebDAV server, checking for errors.
}
