% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav-package.R
\docType{package}
\name{webdav-package}
\alias{webdav}
\alias{webdav-package}
\title{webdav: Access WebDAV servers}
\description{
Read and write from WebDAV servers over http.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxheld83/webdav}
  \item Report bugs at \url{https://github.com/maxheld83/webdav/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
}

}
\keyword{internal}
