% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_copy_file}
\alias{webdav_copy_file}
\title{Copy a resource on the OwnCloud server using WebDAV}
\usage{
webdav_copy_file(
  base_url,
  source_path,
  destination_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD")
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "_server_url_").}

\item{source_path}{The source path of the resource to copy.}

\item{destination_path}{The destination path where the resource will be copied.}

\item{dav}{The WebDAV base path (e.g., "_dav_").}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}
}
\value{
Logical value indicating whether the resource was copied successfully.
}
\description{
This function copies a resource from one URI to another on the OwnCloud server using the WebDAV COPY method.
}
\examples{
\dontrun{
webdav_copy_file("_server_url_", "_file_", "_resource_", "_dav_")
}
}
