% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_create_directory}
\alias{webdav_create_directory}
\title{Create a collection (directory) on the OwnCloud server using WebDAV}
\usage{
webdav_create_directory(
  base_url,
  folder_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD")
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "https://drive.expresso.pe.gov.br").}

\item{folder_path}{The path of the directory to create.}

\item{dav}{The WebDAV base path (e.g., "0c75a584-017d-103a-9c84-d34d2e44200a").}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}
}
\value{
Logical value indicating whether the collection was created successfully.
}
\description{
This function creates a collection (directory) on the OwnCloud server using the WebDAV MKCOL method.
}
\examples{
\dontrun{
webdav_create_directory("https://drive.expresso.pe.gov.br", "/Shared/der/app_painel/new_folder/", "0c75a584-017d-103a-9c84-d34d2e44200a")
}
}
