% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_create_request}
\alias{webdav_create_request}
\title{Create a WebDAV request for the OwnCloud server}
\usage{
webdav_create_request(
  base_url,
  path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD"),
  verbose = FALSE
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "https://drive.expresso.pe.gov.br").}

\item{path}{The specific path for the resource in the OwnCloud WebDAV server.}

\item{dav}{The WebDAV base path (e.g., "0c75a584-017d-103a-9c84-d34d2e44200a").}

\item{username}{The username for OwnCloud authentication. Defaults to the "webdav_USERNAME" environment variable.}

\item{password}{The password for OwnCloud authentication. Defaults to the "webdav_PASSWORD" environment variable.}

\item{verbose}{A logical value indicating whether to print detailed debug messages. When set to TRUE, the function will output additional information about its progress and actions. Default is FALSE.}
}
\value{
An `httr2_request` object with authentication and base URL configured.
}
\description{
This function creates a base WebDAV request for the OwnCloud server with the proper authentication.
}
\examples{
webdav_create_request("https://drive.expresso.pe.gov.br", "/Shared/der/app_painel/data/", "0c75a584-017d-103a-9c84-d34d2e44200a")

}
