% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_delete_resource}
\alias{webdav_delete_resource}
\title{Delete a file or directory from the OwnCloud server using WebDAV}
\usage{
webdav_delete_resource(
  base_url,
  resource_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD"),
  verbose = FALSE
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "https://drive.expresso.pe.gov.br").}

\item{resource_path}{The path of the file or directory to delete.}

\item{dav}{The WebDAV base path (e.g., "0c75a584-017d-103a-9c84-d34d2e44200a").}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}
}
\value{
Logical value indicating whether the file or directory was deleted successfully.
}
\description{
This function deletes a file or directory on the OwnCloud server using the WebDAV DELETE method.
}
\examples{
\dontrun{
webdav_delete_resource("https://drive.expresso.pe.gov.br", "/Shared/der/app_painel/data/file.txt", "0c75a584-017d-103a-9c84-d34d2e44200a")
}
}
