% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_list_files}
\alias{webdav_list_files}
\title{List files from a specific folder on OwnCloud server using WebDAV}
\usage{
webdav_list_files(
  base_url,
  folder_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD"),
  depth = 1,
  verbose = FALSE
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "https://drive.expresso.pe.gov.br").}

\item{folder_path}{The path inside OwnCloud where the files are located.}

\item{dav}{The WebDAV base path (e.g., "0c75a584-017d-103a-9c84-d34d2e44200a").}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}

\item{depth}{The depth of the PROPFIND request (default is 1).}

\item{verbose}{A logical value indicating whether to print detailed debug messages. When set to TRUE, the function will output additional information about its progress and actions. Default is FALSE.}
}
\value{
A tibble with the file names and paths relative to the folder.
}
\description{
This function lists the files in a specific folder on the OwnCloud server using WebDAV, with authentication and error handling.
}
\examples{
webdav_list_files("https://drive.expresso.pe.gov.br", "/Shared/der/app_painel/data/", "0c75a584-017d-103a-9c84-d34d2e44200a")

}
