% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_upload_file}
\alias{webdav_upload_file}
\title{Upload a file to the OwnCloud server using WebDAV}
\usage{
webdav_upload_file(
  base_url,
  file_path,
  upload_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD"),
  timeout = 300,
  verbose = FALSE
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "_server_url_").}

\item{file_path}{The local path of the file to be uploaded.}

\item{upload_path}{The folder path on the OwnCloud server where the file will be uploaded.}

\item{dav}{The WebDAV base path (e.g., "_dav_").}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}

\item{timeout}{The timeout for the upload request in seconds (default is 300 seconds).}

\item{verbose}{A logical value indicating whether to print detailed debug messages. When set to TRUE, the function will output additional information about its progress and actions. Default is FALSE.}
}
\value{
Logical value indicating whether the file was uploaded successfully.
}
\description{
This function uploads a file to a specific folder on the OwnCloud server using WebDAV, with authentication and error handling.
}
\examples{
\dontrun{
webdav_upload_file("_server_url_", "_file_", "_folder_", "_dav_")
}

}
