% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_load_package}
\alias{check_and_load_package}
\title{Check if a package is installed and load it}
\usage{
check_and_load_package(package_name)
}
\arguments{
\item{package_name}{A string with the name of the package to check and load.}
}
\value{
Invisibly returns `TRUE` if the package is successfully loaded or installed and loaded.
If the installation or loading fails, an error will be raised.
}
\description{
This function checks if a specified package is installed in the R environment.
If the package is not installed, it will be automatically installed. After
installation (if necessary), the package is loaded into the session.
}
\examples{
check_and_load_package("httr2")
check_and_load_package("xml2")
}
