% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_delete_resource}
\alias{webdav_delete_resource}
\title{Delete a file or directory from the OwnCloud server using WebDAV}
\usage{
webdav_delete_resource(
  base_url,
  resource_path,
  dav,
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD"),
  verbose = FALSE
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "_server_url_").}

\item{resource_path}{The path of the file or directory to delete.}

\item{dav}{The WebDAV base path (e.g., "_dav_").}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}

\item{verbose}{A logical value indicating whether to print detailed debug messages. When set to TRUE, the function will output additional information about its progress and actions. Default is FALSE.}
}
\value{
Logical value indicating whether the file or directory was deleted successfully.
}
\description{
This function deletes a file or directory on the OwnCloud server using the WebDAV DELETE method.
}
\examples{
\dontrun{
webdav_delete_resource("_server_url_", "_file_", "_dav_")
}
}
