% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{webdav_lock_resource}
\alias{webdav_lock_resource}
\title{Lock a resource on the OwnCloud server using WebDAV}
\usage{
webdav_lock_resource(
  base_url,
  resource_path,
  dav,
  lock_type = "exclusive",
  username = Sys.getenv("OWNCLOUD_USERNAME"),
  password = Sys.getenv("OWNCLOUD_PASSWORD")
)
}
\arguments{
\item{base_url}{The base URL of the OwnCloud server (e.g., "https://drive.expresso.pe.gov.br").}

\item{resource_path}{The path of the resource to lock.}

\item{dav}{The WebDAV base path (e.g., "0c75a584-017d-103a-9c84-d34d2e44200a").}

\item{lock_type}{Type of lock ('exclusive' or 'shared').}

\item{username}{The username for OwnCloud authentication.}

\item{password}{The password for OwnCloud authentication.}
}
\value{
Logical value indicating whether the resource was locked successfully.
}
\description{
This function locks a resource on the OwnCloud server using the WebDAV LOCK method.
}
\examples{
\dontrun{
webdav_lock_resource("https://drive.expresso.pe.gov.br", "/Shared/der/app_painel/data/file.txt", "0c75a584-017d-103a-9c84-d34d2e44200a", "exclusive")
}
}
