% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav_put.R
\name{webdav_put}
\alias{webdav_put}
\title{PUT new file on WebDAV server}
\usage{
webdav_put(local, url, user)
}
\arguments{
\item{local}{\code{character[1]}, giving the path to the local file to upload.}

\item{url}{\code{character[1]}, giving the path to the file to create or overwrite on the WebDAV server.
Remember to include the target filename.}

\item{user}{\code{character[1]}, giving the username to log in with.}
}
\value{
\link[httr:response]{httr::response}
}
\description{
PUTs new file into a WebDAV drive, in the HTTP sense:
Existing files at the same \code{url} are overwritten.
Will retry as per \code{\link[httr:retry]{httr::retry()}}.
}
